Refactors ARTIS reader into dedicated module structure
Reorganizes ARTIS-related code by moving it from a single reader file into a dedicated module with separate data and reader components.

Extracts the ArtisModelData class into a standalone data module and renames it to ArtisData for consistency.

Improves code organization and maintainability by separating data structures from reading functionality.
