+++ tardis/io/model/artis/readers.py
@@ -1,34 +1,9 @@
-from dataclasses import dataclass, field
 
 import numpy as np
 import pandas as pd
 from astropy import units as u
 
-from tardis.model.geometry.radial1d import HomologousRadial1DGeometry
-
-
-@dataclass
-class ArtisModelData:
-    time_of_model: u.Quantity
-    velocity: np.ndarray
-    mean_density: u.Quantity
-    mass_fractions: pd.DataFrame = field(default_factory=pd.DataFrame)
-
-    def to_geometry(self):
-        """
-        Construct a HomologousRadial1DGeometry object from this ArtisModelData.
-
-        We create v_inner and v_outer by treating the velocity array as boundary
-        points for the shells. The time_of_model is used as the time_explosion.
-        """
-        geometry = HomologousRadial1DGeometry(
-            v_inner=self.velocity[:-1],
-            v_outer=self.velocity[1:],
-            v_inner_boundary=None,
-            v_outer_boundary=None,
-            time_explosion=self.time_of_model,
-        )
-        return geometry
+from tardis.io.model.artis.data import ArtisData
 
 
 def read_artis_density(fname, legacy_return=True):
@@ -188,7 +163,7 @@ def read_artis_model(density_fname, abundance_fname):
 
     mass_fractions = pd.concat([mass_fractions, isotope_mass_fractions], axis=0)
 
-    return ArtisModelData(
+    return ArtisData(
         time_of_model=time_of_model,
         velocity=velocity,
         mean_density=mean_density,
